Move query parser tests from Solidity into testlang
