+++ crates/solidity/outputs/cargo/tests/src/lib.rs
@@ -8,5 +8,4 @@ mod bindings_rules;
 mod cst_output;
 mod doc_examples;
 mod generated;
-mod query;
 mod trivia;
