--- crates/solidity/outputs/cargo/tests/src/query/mod.rs
@@ -1,1 +1,0 @@
-mod parser_tests;
